+++ pypeec/lib_mesher/mesher_stl.py
@@ -66,7 +66,7 @@ def _get_voxelize_stl(grid, mesh):
         # get the indices of the extracted voxels
         idx_voxel = voxel["idx"]
     else:
-        idx_voxel = np.empty(0, dtype=np.int_)
+        idx_voxel = np.empty(0, dtype=np.int64)
 
     return idx_voxel
 
@@ -82,8 +82,8 @@ def _get_mesh_stl(domain_stl):
     mesh_all = []
 
     # init the coordinate (minimum and maximum coordinates)
-    xyz_min = np.full(3, +np.inf, dtype=np.float_)
-    xyz_max = np.full(3, -np.inf, dtype=np.float_)
+    xyz_min = np.full(3, +np.inf, dtype=np.float64)
+    xyz_max = np.full(3, -np.inf, dtype=np.float64)
 
     # load the STL files and find the bounding box
     for tag, domain_stl_tmp in domain_stl.items():
@@ -98,8 +98,8 @@ def _get_mesh_stl(domain_stl):
 
             # find the bounds
             (x_min, x_max, y_min, y_max, z_min, z_max) = mesh.bounds
-            tmp_min = np.array((x_min, y_min, z_min), dtype=np.float_)
-            tmp_max = np.array((x_max, y_max, z_max), dtype=np.float_)
+            tmp_min = np.array((x_min, y_min, z_min), dtype=np.float64)
+            tmp_max = np.array((x_max, y_max, z_max), dtype=np.float64)
 
             # update the bounds
             xyz_min = np.minimum(xyz_min, tmp_min)
@@ -129,8 +129,8 @@ def _get_voxel_size(d, xyz_max, xyz_min):
     d = (xyz_max-xyz_min)/n
 
     # cast data
-    d = d.astype(np.float_)
-    n = n.astype(np.int_)
+    d = d.astype(np.float64)
+    n = n.astype(np.int64)
 
     # check voxel validity
     if not np.all(d > 0):
@@ -157,7 +157,7 @@ def _get_voxel_grid(n, d, c):
     grid.spacing = d
 
     # add indices for tracking the voxels after voxelization
-    grid["idx"] = np.arange(np.prod(n), dtype=np.int_)
+    grid["idx"] = np.arange(np.prod(n), dtype=np.int64)
 
     # cast is required for voxelization
     grid = grid.cast_to_unstructured_grid()
@@ -173,7 +173,7 @@ def _get_domain_def(grid, domain_stl, mesh_stl):
     # init the domain dict
     domain_def = {}
     for tag in domain_stl:
-        domain_def[tag] = np.empty(0, np.int_)
+        domain_def[tag] = np.empty(0, np.int64)
 
     # voxelize the meshes
     for mesh_stl_tmp in mesh_stl:
@@ -213,11 +213,11 @@ def get_mesh(param, domain_stl):
 
     # if provided, the specified bounds are used, otherwise the STL bounds are used
     if xyz_min is not None:
-        xyz_min = np.array(xyz_min, np.float_)
+        xyz_min = np.array(xyz_min, np.float64)
     else:
         xyz_min = xyz_min_stl
     if xyz_max is not None:
-        xyz_max = np.array(xyz_max, np.float_)
+        xyz_max = np.array(xyz_max, np.float64)
     else:
         xyz_max = xyz_max_stl
 
@@ -240,8 +240,8 @@ def get_mesh(param, domain_stl):
 
     # cast reference mesh
     reference = {
-        "faces": np.array(reference.faces, dtype=np.int_),
-        "points": np.array(reference.points, dtype=np.float_),
+        "faces": np.array(reference.faces, dtype=np.int64),
+        "points": np.array(reference.points, dtype=np.float64),
     }
 
     return n, d, c, domain_def, reference
